mise a jour du module poco

+++ app/cpp/qmake/mingw/code/src/manager/GCode.cpp
@@ -149,7 +149,7 @@ bool GCode::addData(const GString& _code, const std::vector<GLog*>& _map) {
         createNode("map");
         next();
     }
-    for(int i = 0; i < _map.size(); i++) {
+    for(int i = 0; i < (int)_map.size(); i++) {
         GLog* lObj = _map.at(i);
         GString lData = lObj->serialize();
         GCode lDom;
